Introduce staking RPC module
